Implementing a Logger interface

+++ ContonsoUniversity/Models/Course.cs
@@ -9,7 +9,7 @@ namespace ContonsoUniversity.Models
     public class Course
     {
         [DatabaseGenerated(DatabaseGeneratedOption.None)]       
-        public Int64 CourseID { get; set; }
+        public int CourseID { get; set; }
         public string Title { get; set; }
         public int Credits { get; set; }
         public virtual ICollection<Enrollment> Enrollments { get; set; }
